renamed CUDA mem backend header
